Move ps2000_wrapper_blockmode_utils to instruments folder
